+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Additions
 
 Improvements
 - Allow elements to auto-adjust their size even when their children are aligned oddly
+- Close other dropdowns when opening a dropdown
 
 ## 6.0.0
 ### MLEM